consolidate oper12_view and oper12_row_view due to tight coupling
deleted:    include/oa/accel/oper12_row_view.h
modified:   include/oa/accel/oper12_view.h
modified:   src/accel/CMakeLists.txt
deleted:    src/accel/oper12_row_view.cpp
modified:   src/accel/oper12_view.cpp
