+++ src/accel/CMakeLists.txt
@@ -9,7 +9,6 @@ if(XLLSDK_FOUND)
         fp12_view.cpp
         mref12.cpp
         oper12.cpp
-        oper12_row_view.cpp
         oper12_view.cpp
         udf.cpp
         xl_conv.cpp
